requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
httpx==0.28.1
supabase==2.11.0
pydantic-settings==2.7.1
PyJWT==2.10.1
